% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krikinton-package.R
\docType{package}
\name{krikinton}
\alias{krikinton}
\alias{krikinton-package}
\title{krikinton: rJava Wrapper of Kintoki}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

rJava wrapper of Kintoki (a port of the CaboCha to Java) and Sudachi.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/krikinton}
  \item Report bugs at \url{https://github.com/paithiov909/krikinton/issues}
}

}
\author{
\strong{Maintainer}: Akiru Kato \email{paithiov909@gmail.com}

}
\keyword{internal}
