% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudachi.R
\name{rebuild_dictionary}
\alias{rebuild_dictionary}
\title{Rebuild Dictionary}
\usage{
rebuild_dictionary(json_string)
}
\arguments{
\item{json_string}{character scalar. valid json string}
}
\value{
returns a binding to the instance of
`com.worksap.nlp.sudachi.JapaneseDictionary` invisibly
}
\description{
Rebuild Dictionary
}
