% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intarface-kintoki.R
\name{rebuild_parser}
\alias{rebuild_parser}
\title{Rebuild Kintoki Parser}
\usage{
rebuild_parser()
}
\value{
returns a binding to the instance of
\code{com.worksap.nlp.kintoki.cabocha.Cabocha} invisibly
}
\description{
Rebuild Kintoki Parser
}
