% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-sudachi.R
\name{rebuild_tokenizer}
\alias{rebuild_tokenizer}
\title{Rebuild Sudachi Tokenizer}
\usage{
rebuild_tokenizer()
}
\value{
a binding to the instance of
\code{com.worksap.nlp.sudachi.JapaneseTokenizer} is returned invisibly.
}
\description{
Rebuild Sudachi Tokenizer
}
