% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-string.R
\docType{data}
\name{sudachi_json}
\alias{sudachi_json}
\title{List Presentation of Sudachi's Setting}
\format{
An object of class \code{list} of length 5.
}
\usage{
sudachi_json
}
\description{
List Presentation of Sudachi's Setting
}
\keyword{datasets}
