% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krikinton-pacakge.R
\docType{package}
\name{krikinton}
\alias{krikinton}
\alias{krikinton-package}
\title{krikinton: rJava Wrapper of Kintoki}
\description{
rJava wrapper of WorksApplications/kintoki
  (a port of the CaboCha to Java).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paithiov909/krikinton}
  \item Report bugs at \url{https://github.com/paithiov909/krikinton/issues}
}

}
\author{
\strong{Maintainer}: Kato Akiru \email{paithiov909@gmail.com}

}
\keyword{internal}
