% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sudachi.R
\name{rebuild_tokenizer}
\alias{rebuild_tokenizer}
\title{Rebuild Sudachi Tokenizer}
\usage{
rebuild_tokenizer()
}
\value{
returns a binding to the instance of
`tokenizer` invisibly
}
\description{
Rebuild Sudachi Tokenizer
}
